+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Gallery item captions are no longer wrapped in a `<p>` element
 - Updated the heading link arrow style
 - Added more space below table elements
+- Removed focus shift on accordion items upon opening
 
 ### Deprecated
 ### Removed